timeout fix

+++ signals.cpp
@@ -44,19 +44,22 @@ void alarmHandler(int sig_num, siginfo_t* info, void* context) {
 	SmallShell& smash = SmallShell::getInstance();
 	std::string cmd = smash.timed_list.front().timeout_cmd;
 	int pid = smash.timed_list.front().pid_cmd;
+	int alrm_time = smash.timed_list.front().alrm_time;
 	std::string str("smash: ");
 	str.append(cmd).append(" timed out!\n");
 
-	if (kill(pid, SIGKILL) == ERROR)
+	if (alrm_time != EXITED)
 	{
-		fprintf(stderr, "smash error: kill failed\n");
-		return;
+		if (kill(pid, SIGKILL) == ERROR)
+		{
+			fprintf(stderr, "smash error: kill failed\n");
+			return;
+		}
+		cout << str;
 	}
 	smash.timed_list.pop_front();
 	smash.timed_list.sort();
 	if (!smash.timed_list.empty())
 		alarm(difftime(smash.timed_list.front().alrm_time, time(NULL)));
-	
-	cout << str;
 }
